+++ NEWS.md
@@ -1,9 +1,10 @@
 # plotBart 0.1.4
 - propensity_scores() no longer exported
+- Removed legend argument from `plot_moderator_c_pd()`. Legend can be modified using `p + theme(legend.position = "<position>")`
 
 # plotBart 0.1.3
 - Homogenized function argument names across functions. This is code breaking as previous argument names are not supported.
-- `plot_moderator_search()` now returns a ggplot object instead of an rpart.plot object
+- `plot_moderator_search()` now returns a ggplot object instead of an `rpart.plot` object
 
 # plotBart 0.1.2
 - Fixed bug in waterfall plot